+++ src/main/java/kz/syllabus/repository/PositionRepository.java
@@ -1,10 +1,9 @@
 package kz.syllabus.repository;
 
 import kz.syllabus.entity.Position;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PositionRepository extends JpaRepository<Position, Integer> {
-
-}
+public interface PositionRepository extends JpaRepository<Position, Long> {}
