+++ src/main/java/kz/syllabus/repository/ProgramDetailRepository.java
@@ -1,11 +1,13 @@
 package kz.syllabus.repository;
 
 import kz.syllabus.entity.ProgramDetail;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProgramDetailRepository extends JpaRepository<ProgramDetail, Integer> {
-    ProgramDetail getBySyllabusProgramId(Integer syllabusProgramId);
-    ProgramDetail getProgramDetailBySyllabusProgramId(Integer id);
+public interface ProgramDetailRepository extends JpaRepository<ProgramDetail, Long> {
+    ProgramDetail getBySyllabusProgramId(Long syllabusProgramId);
+
+    ProgramDetail getProgramDetailBySyllabusProgramId(Long id);
 }
