+++ src/main/java/kz/syllabus/repository/RoleRepository.java
@@ -1,13 +1,11 @@
 package kz.syllabus.repository;
 
+import kz.syllabus.entity.user.Role;
 
-import kz.syllabus.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-//    Optional<Role> findByName(String name);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    //    Optional<Role> findByName(String name);
 }
