+++ src/main/java/kz/syllabus/repository/SyllabusProgramRepository.java
@@ -1,13 +1,15 @@
 package kz.syllabus.repository;
 
-import kz.syllabus.entity.SyllabusProgram;
+import kz.syllabus.entity.syllabus.SyllabusProgram;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface SyllabusProgramRepository extends JpaRepository<SyllabusProgram, Integer> {
-    SyllabusProgram getBySyllabusId(Integer disciplineInfoId);
-    List<SyllabusProgram> getAllBySyllabusId(Integer syllabusId);
+public interface SyllabusProgramRepository extends JpaRepository<SyllabusProgram, Long> {
+    SyllabusProgram getBySyllabusId(Long disciplineInfoId);
+
+    List<SyllabusProgram> getAllBySyllabusId(Long syllabusId);
 }
