+++ src/main/java/kz/syllabus/repository/TestUserRepository.java
@@ -1,10 +1,11 @@
 package kz.syllabus.repository;
 
-import kz.syllabus.entity.TestUser;
+import kz.syllabus.entity.user.TestUser;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestUserRepository extends JpaRepository<TestUser, Integer> {
+public interface TestUserRepository extends JpaRepository<TestUser, Long> {
     TestUser getByIin(String iin);
 }
